requirements.txt
flask=<4.0
PyMuPDF=<2.0
pdf2docx=<1.0
sumy=<1.0
nltk=<4.0
gunicorn=<23.0
requests=<3.0
python-dotenv=<2.0
